Autoscaler should take active tasks into account
It seems like the autoscaler only considered
reserved tasks when deciding to scale the pool.
This is sub-optimal, consider the following:

1 current active task
1 reserved task
1 current process

The autoscaler would compare the reserved (1)
tasks with the current number of processes (1)
and decide not to scale up. However, this means
that the reserved task would have to wait for
the active one to finish even though it is
allowed to launch more processes.